fix players and clubs

+++ LigaNOS/Controllers/ClubsController.cs
@@ -30,7 +30,7 @@ namespace LigaNOS.Controllers
         // GET: ClubsController
         public IActionResult Index()
         {
-            return View(_clubRepository.GetAllWithClubs().OrderBy(c => c.Name));
+            return View(_clubRepository.GetAll().OrderBy(c => c.Name));
         }
 
         // GET: ClubsController/Details/5
@@ -73,9 +73,14 @@ namespace LigaNOS.Controllers
 
                 }
                 var club = _converterHelper.ToClub(model, imageId, true);
+                
+                if(!User.Identity.IsAuthenticated)
+                {
+                    return RedirectToAction("Login", "Account");
+                }
 
                 club.User = await _userHelper.GetUserByEmailAsync(this.User.Identity.Name);
-                await _clubRepository.UpdateAsync(club);
+               
 
                 await _clubRepository.CreateAsync(club);
                 return RedirectToAction(nameof(Index));
@@ -84,14 +89,10 @@ namespace LigaNOS.Controllers
         }
 
         // GET: ClubsController/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
+        public async Task<IActionResult> Edit(int id)
+        { 
 
-            var club = await _clubRepository.GetByIdAsync(id.Value);
+            var club = await _clubRepository.GetByIdAsync(id);
             if (club == null)
             {
                 return NotFound();
@@ -104,22 +105,41 @@ namespace LigaNOS.Controllers
         // POST: ClubsController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(ClubViewModel model)
+        public async Task<IActionResult> Edit(int id, ClubViewModel model)
         {
+            if (id != model.Id)
+            {
+                return NotFound();
+            }
+
             if (ModelState.IsValid)
             {
+                if (!User.Identity.IsAuthenticated)
+                {
+                    return RedirectToAction("Login", "Account");
+                }
+
                 try
                 {
-                    Guid imageId = Guid.Empty;
-
+                    var club = await _clubRepository.GetByIdAsync(model.Id);
+                    if (club == null)
+                    {
+                        return NotFound();
+                    }
+                    // Check if a new image file is provided
                     if (model.ImageFile != null && model.ImageFile.Length > 0)
                     {
-
-                        imageId = await _blobHelper.UploadBlobAsync(model.ImageFile, "clubs");
+                        club.ImageFileId = await _blobHelper.UploadBlobAsync(model.ImageFile, "clubs");
                     }
-                    var club = _converterHelper.ToClub(model, imageId, false);
 
-                    club.User = await _userHelper.GetUserByEmailAsync(this.User.Identity.Name);
+                    
+                    club.Name = model.Name;
+                    club.Coach = model.Coach;
+                    club.Stadium = model.Stadium;
+      
+                    club.User = await _userHelper.GetUserByEmailAsync(User.Identity.Name);
+
+                    await _clubRepository.UpdateAsync(club);
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -133,9 +153,9 @@ namespace LigaNOS.Controllers
                     }
                 }
                 return RedirectToAction(nameof(Index));
-                }
-                return View(model);
             }
+            return View(model);
+        }
 
             // GET: ClubsController/Delete/5
             public async Task<IActionResult> Delete(int? id)
@@ -160,16 +180,24 @@ namespace LigaNOS.Controllers
             [ValidateAntiForgeryToken]
             public async Task<IActionResult> DeleteConfirmed(int id)
             {
-                var club = await _clubRepository.GetByIdAsync(id);
-                if (club == null)
-                {
-                    return NotFound();
-                }
-
-                await _clubRepository.DeleteAsync(club);
-                return RedirectToAction(nameof(Index));
+            var club = await _clubRepository.GetByIdAsync(id);
+            if (club == null)
+            {
+                return NotFound();
             }
+
+            var hasMatches = await _clubRepository.HasMatchesAsync(id); // Use repository method
+
+            if (hasMatches)
+            {
+                ModelState.AddModelError(string.Empty, "Cannot delete this club because it has matches associated.");
+                return View(club);
+            }
+
+            await _clubRepository.DeleteAsync(club);
+            return RedirectToAction(nameof(Index));
         }
     }
+}
